requirements.txt
anytree=2.6.0
mysql.connector=2.2.9